requirements.txt
annotated-types
anyio
boto3
botocore
certifi
charset-normalizer
click
dnspython
email-validator
fastapi
fastapi-cli
fastapi-cloud-cli
filelock
fsspec
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
idna
Jinja2
jmespath
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
numpy
packaging
pydantic
pydantic_core
Pygments
python-dateutil
python-dotenv
python-multipart
PyYAML
regex
requests
rich
rich-toolkit
rignore
s3transfer
safetensors
sentencepiece
sentry-sdk
setuptools
shellingham
six
sniffio
starlette
sympy
tokenizers
torch
tqdm
transformers
typer
typing-inspection
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
openai
aiohttp
nest-asyncio
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-experimental
langchain-huggingface
langchain-mcp-adapters
langsmith
# Autenticación
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
pyjwt
